requirements.txt
admin==0.0.1
asgiref==3.3.1
Django==3.1.6
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.2
Faker==6.1.1
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
zipp==3.4.0
